backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
django-cors-headers>=4.3
openai>=1.0
anthropic>=0.40
gunicorn>=23.0
whitenoise>=6.8
python-dotenv>=1.0
